+++ component/pl_mdns_client.cpp
@@ -1,4 +1,5 @@
 #include "pl_mdns_client.h"
+#include "pl_mdns_base.h"
 #include "mdns.h"
 #include "esp_check.h"
 
@@ -34,8 +35,13 @@ esp_err_t MdnsClient::DnsSdQuery(const std::string& type, const std::string& pro
   mdns_result_t* results = NULL;
   serviceInstancesInfo.clear();
 
-  ESP_RETURN_ON_ERROR(mdns_init(), TAG, "init failed");
-  ESP_RETURN_ON_ERROR(mdns_query(NULL, type.c_str(), protocol.c_str(), MDNS_TYPE_PTR, readTimeout * portTICK_PERIOD_MS, maxNumberOfInstances, &results), TAG, "query failed");
+  ESP_RETURN_ON_ERROR(Mdns::Init(), TAG, "init failed");
+
+  esp_err_t error = mdns_query(NULL, type.c_str(), protocol.c_str(), MDNS_TYPE_PTR, readTimeout * portTICK_PERIOD_MS, maxNumberOfInstances, &results);
+  if (error != ESP_OK)
+    Mdns::Free();
+  ESP_RETURN_ON_ERROR(error, TAG, "query failed");
+
   if (results) {
     for (mdns_result_t* r = results; r; r = r->next) {
       MdnsServiceInstanceInfo info;
@@ -62,7 +68,8 @@ esp_err_t MdnsClient::DnsSdQuery(const std::string& type, const std::string& pro
     mdns_query_results_free(results);
   }
 
-  return ESP_OK;  
+  Mdns::Free();
+  return ESP_OK;
 }
 
 //==============================================================================